+++ CHANGES.md
@@ -2,10 +2,21 @@
 
 ## TODO
 
-- [TODO] Support for reading config from EDN files
-- [TODO] Uniform interface for resolving config from various kinds of files (JSON, YAML etc.)
-- [TODO] Hierarchical config resolution support for all types of config files
 - [TODO] Export specified key/value pairs as system properties
+
+
+## 0.4.0 / 2016-June-??
+
+- Config files support
+  - Support for reading config from EDN files
+  - Unified interface for resolving config from various kinds of files (JSON, YAML etc.)
+    - Hierarchical config resolution support for all types of config files
+  - Support for writing config files
+- Deprecated API
+  - [TODO] Deprecate class `keypin.PropertyFile`
+  - [TODO] Deprecate `keypin.core/read-properties` and `keypin.core/lookup-property`
+- Dependency injection
+  - [TODO] Helpers to partially apply annotated arguments in a function
 
 
 ## 0.3.2 / 2016-June-03